Update ConsumptionHistorySource.scala
@patsonluk 
Ok this is my test code that super optimizes this by saving the consumption history as a single SQL chunk instead of thousands (and thousands) of separate SQL inserts. Had to create a new table for this, felt you should take a look and implement it as you see fit. Saved 150-200 seconds of time from every cycle for me