Update AirportFeaturePatcher.scala
Hi patson! Changes proposed are all cosmetic at this stage. No new charm cities have been added and no charm values adjusted in this change. I will propose charm additions on next commit if this change is accepted. This is just a code cleanup to make it easy to see which airport/cities have charms and the value of those charms.

Changes made are:
1. Added comments detailing all Airport Names for all Charms so airport location is easily recognizable
2. Fixed capitalization of all Airport Names in comments of Charms 
3. Sorted airport by Charm value from highest to lowest