Added Xian MA60, Chinese aircraft based on An-24
The Xian MA60 is a Chinese aircraft based on the Xian Y-7, which is based on the An-24. Many webpages say that it has STOL capabilities. I used the An-24's runway req for this aircraft, found at https://www.deagel.com/Airliners/An-24/a002598
I put manufacturer as Comac because making a new one is problematic
Stats: 62 cap, 2.7 fuel burn, 430 speed, 1600 range, costs 20mil, 25 year lifespan, runway requirement 620m